@@ -1,21 +1,25 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.4.20'
     ext.lifecycle_version = '1.1.1'
     ext.retrofit_version = '2.4.0'
     ext.dagger2_version = '2.16'
     ext.android_support_version = '28.0.0'
     ext.room_version = '1.1.1'
     ext.gradle_version = '3.1.0'
+    ext.hilt_version = "2.28-alpha"
+    ext.navigation_version = '2.2.2'
 
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:3.2.0"
+        classpath 'com.android.tools.build:gradle:4.1.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$navigation_version"
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
