@@ -1,42 +1,39 @@
 package todo.quang.mvvm.ui.post
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.annotation.StringRes
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
+import androidx.annotation.StringRes
+import com.google.android.material.snackbar.Snackbar
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
+import android.util.Log
+import androidx.activity.viewModels
+import dagger.hilt.android.AndroidEntryPoint
 import todo.quang.mvvm.R
 import todo.quang.mvvm.databinding.ActivityPostListBinding
-import todo.quang.mvvm.injection.ViewModelFactory
 
-class PostListActivity: AppCompatActivity() {
+@AndroidEntryPoint
+class PostListActivity : AppCompatActivity() {
     private lateinit var binding: ActivityPostListBinding
-    private lateinit var viewModel: PostListViewModel
     private var errorSnackbar: Snackbar? = null
+    private val viewModel: PostListViewModel by viewModels()
 
-    override fun onCreate(savedInstanceState: Bundle?){
+    override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         binding = DataBindingUtil.setContentView(this, R.layout.activity_post_list)
         binding.postList.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
-
-        viewModel = ViewModelProviders.of(this, ViewModelFactory(this)).get(PostListViewModel::class.java)
-        viewModel.errorMessage.observe(this, Observer {
-            errorMessage -> if(errorMessage != null) showError(errorMessage) else hideError()
+        viewModel.genreLiveData.observe(this, {
+            Log.d("groupApp", "onCreate: ${it?.data?.get(0)}")
         })
-        binding.viewModel = viewModel
     }
 
-    private fun showError(@StringRes errorMessage:Int){
+    private fun showError(@StringRes errorMessage: Int) {
         errorSnackbar = Snackbar.make(binding.root, errorMessage, Snackbar.LENGTH_INDEFINITE)
-        errorSnackbar?.setAction(R.string.retry, viewModel.errorClickListener)
         errorSnackbar?.show()
     }
 
-    private fun hideError(){
+    private fun hideError() {
         errorSnackbar?.dismiss()
     }
 }
