@@ -1,15 +1,15 @@
 package todo.quang.mvvm.ui.post
 
-import android.databinding.DataBindingUtil
-import android.support.v7.widget.RecyclerView
+import androidx.databinding.DataBindingUtil
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import todo.quang.mvvm.R
 import todo.quang.mvvm.databinding.ItemPostBinding
-import todo.quang.mvvm.model.Post
+import todo.quang.mvvm.model.AppInfo
 
 class PostListAdapter: RecyclerView.Adapter<PostListAdapter.ViewHolder>() {
-    private lateinit var postList:List<Post>
+    private lateinit var appInfoList:List<AppInfo>
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding: ItemPostBinding = DataBindingUtil.inflate(LayoutInflater.from(parent.context), R.layout.item_post, parent, false)
@@ -17,24 +17,17 @@ class PostListAdapter: RecyclerView.Adapter<PostListAdapter.ViewHolder>() {
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(postList[position])
     }
 
     override fun getItemCount(): Int {
-        return if(::postList.isInitialized) postList.size else 0
+        return if(::appInfoList.isInitialized) appInfoList.size else 0
     }
 
-    fun updatePostList(postList:List<Post>){
-        this.postList = postList
+    fun updatePostList(appInfoList:List<AppInfo>){
+        this.appInfoList = appInfoList
         notifyDataSetChanged()
     }
 
-    class ViewHolder(private val binding: ItemPostBinding):RecyclerView.ViewHolder(binding.root){
-        private val viewModel = PostViewModel()
-
-        fun bind(post: Post){
-            viewModel.bind(post)
-            binding.viewModel = viewModel
-        }
+    class ViewHolder(private val binding: ItemPostBinding): RecyclerView.ViewHolder(binding.root){
     }
 }
