@@ -1,7 +1,7 @@
 package todo.quang.mvvm.utils.extension
 
 import android.content.ContextWrapper
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 
 fun View.getParentActivity(): AppCompatActivity?{
