@@ -1,11 +1,15 @@
 package todo.quang.mvvm.model.database
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
-import todo.quang.mvvm.model.Post
-import todo.quang.mvvm.model.PostDao
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import todo.quang.mvvm.model.AppInfo
+import todo.quang.mvvm.model.AppInfoDao
 
-@Database(entities = [Post::class], version = 1)
+@Database(entities = [AppInfo::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
-    abstract fun postDao(): PostDao
+    abstract fun appInfoDao(): AppInfoDao
+
+    companion object {
+        const val DATABASE_NAME = "app_db"
+    }
 }
