@@ -1,4 +1,4 @@
 package todo.quang.mvvm.utils
 
 /** The base URL of the API */
-const val BASE_URL: String = "https://jsonplaceholder.typicode.com"
+const val BASE_URL: String = "http://10.0.2.2:5000/"
